+++ man/deprecated_functions.R
@@ -1,0 +1,14 @@
+#' Deprecated functions in multid
+#'
+#' These functions still work but will be removed (defunct) in the next version.
+#'
+#' \itemize{
+#'  \item \code{\link{sem_dadas}}: This function is deprecated, and will
+#'  be removed in the next version of this package.
+#' }
+#' \itemize{
+#'  \item \code{\link{ml_dadas}}: This function is deprecated, and will
+#'  be removed in the next version of this package.
+#' }
+#'
+#' @name multid-deprecated