+++ NEWS.md
@@ -5,6 +5,7 @@
 * Added a possibility to obtain bootstrap estimates and percentile confidence intervals for non-scaled parameter estimates in ddsc_ml results-table
 * Added confidence intervals for ddsc_ml results table
 * Added a possibility to bootstrap in ddsc_sem
+* ml_dadas and sem_dadas deprecated (superceded by ddsc_ml and ddsc_sem)
 
 # multid 0.9.0