+++ NEWS.md
@@ -7,6 +7,7 @@
 * Added a possibility to bootstrap in ddsc_sem
 * ml_dadas and sem_dadas deprecated (superceded by ddsc_ml and ddsc_sem)
 * Added plot_ddsc function for directly plotting ddsc_sem results
+* Removed ml_dadas and sem_dadas from README examples. Replaced with ddsc_ml and ddsc_sem
 
 # multid 0.9.0